Newlines and no spaces.
Changed the code to version 11 of the program. pgn is corrected to
remove unnecessary newlines. Additionally, if a move comes immediately
after the turn number, without a space (as there should be), the
program determines if a period or a space has the highest index and
specifies that as the beginning of the move. This is only relevant to
the section of code that deals with recognizing comments.